+++ @navikt/core/react/src/form/Fieldset/fieldset.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta } from "@storybook/react";
 import React from "react";
-import { TextField } from "../../index";
 import { Fieldset } from "../index";
+import TextField from "../textfield/TextField";
 
 export default {
   title: "ds-react/Fieldset",
