+++ @navikt/core/react/src/form/confirmation-panel/ConfirmationPanel.test.tsx
