+++ @navikt/core/react/src/form/confirmation-panel/confirmation-panel.stories.tsx
