+++ @navikt/core/react/src/form/switch/Switch.tsx
