+++ @navikt/core/react/src/form/textfield/TextField.tsx
@@ -1,9 +1,9 @@
 import cl from "clsx";
 import React, { InputHTMLAttributes, forwardRef } from "react";
-import { BodyShort, ErrorMessage, Label } from "../typography";
-import { omit } from "../util";
-import { ReadOnlyIcon } from "./ReadOnlyIcon";
-import { FormFieldProps, useFormField } from "./useFormField";
+import { BodyShort, ErrorMessage, Label } from "../../typography";
+import { omit } from "../../util";
+import { ReadOnlyIcon } from "../ReadOnlyIcon";
+import { FormFieldProps, useFormField } from "../useFormField";
 
 export interface TextFieldProps
   extends FormFieldProps,
